Moves deprecated components to DeprecatedGame
