+++ QuizEngine/DeprecatedGame.swift
@@ -8,6 +8,15 @@
 import Foundation
 
 @available(*, deprecated)
+public protocol Router {
+    associatedtype Question: Hashable
+    associatedtype Answer
+    
+    func routeTo(question: Question, answerCallback: @escaping (Answer) -> Void)
+    func routeTo(result: Result<Question, Answer>)
+}
+
+@available(*, deprecated)
 public final class Game <Question, Answer, R: Router> {
     let flow: Any
