--- QuizEngine/Router.swift
@@ -1,25 +1,0 @@
-//
-//  Router.swift
-//  QuizEngine
-//
-//  Created by Kalil Holanda on 05/06/21.
-//
-
-import Foundation
-
-public protocol QuizDelegate {
-    associatedtype Question: Hashable
-    associatedtype Answer
-    
-    func handle(question: Question, answerCallback: @escaping (Answer) -> Void)
-    func handle(result: Result<Question, Answer>)
-}
-
-@available(*, deprecated)
-public protocol Router {
-    associatedtype Question: Hashable
-    associatedtype Answer
-    
-    func routeTo(question: Question, answerCallback: @escaping (Answer) -> Void)
-    func routeTo(result: Result<Question, Answer>)
-}
